don't update portpool if it is nil
https://github.com/rancher/rancher/issues/10252.When system is
overloaded the metadata call to get all containers may timeout and
return nil portpool.(portpool is calcaluted from all ports of all
containers running in the current environment). In this case we skip
updating portpool and continue, but increase the failure count so that
scheduler can exit when count reaches threhold(5).